# config/openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "catalog-update",
        "version": "1.0",
        "description": "カタログ更新サービス"
    },
    "servers": [
        {
            "url": "/catalog-update"
        }
    ],
    "paths": {
...
        "/region": {
            "post": {
                "tags": [
                    "申請"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {},
                                "example": {}
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "アクター認定申請取得に成功した際のレスポンスコード",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {},
                                    "example": {}
                                }
                            }
                        }
                    }
                }
            }
        },
        "/region/delete": {
            "post": {
                "tags": [
                    "解除申請"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {},
                                "example": {}
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "アクター認定申請取得に成功した際のレスポンスコード",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {},
                                    "example": {}
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}